Reorganize launcher package
